+++ main.go
@@ -53,7 +53,7 @@ type Options struct {
 
 // NewRouter generate new router with swagger. Default to OpenAPI 3.0.0
 func NewRouter[HandlerFunc, Route any](router apirouter.Router[HandlerFunc, Route], options Options) (*Router[HandlerFunc, Route], error) {
-	swagger, err := generateNewValidSwagger(options.Openapi)
+	swagger, err := generateNewValidOpenapi(options.Openapi)
 	if err != nil {
 		return nil, fmt.Errorf("%w: %s", ErrValidatingSwagger, err)
 	}
@@ -104,7 +104,7 @@ func (r Router[HandlerFunc, Route]) SubRouter(router apirouter.Router[HandlerFun
 	}, nil
 }
 
-func generateNewValidSwagger(swagger *openapi3.T) (*openapi3.T, error) {
+func generateNewValidOpenapi(swagger *openapi3.T) (*openapi3.T, error) {
 	if swagger == nil {
 		return nil, fmt.Errorf("swagger is required")
 	}
@@ -128,9 +128,9 @@ func generateNewValidSwagger(swagger *openapi3.T) (*openapi3.T, error) {
 	return swagger, nil
 }
 
-// GenerateAndExposeSwagger creates a /documentation/json route on router and
+// GenerateAndExposeOpenapi creates a /documentation/json route on router and
 // expose the generated swagger
-func (r Router[_, _]) GenerateAndExposeSwagger() error {
+func (r Router[_, _]) GenerateAndExposeOpenapi() error {
 	if err := r.swaggerSchema.Validate(r.context); err != nil {
 		return fmt.Errorf("%w: %s", ErrValidatingSwagger, err)
 	}