+++ support/echo/echo_test.go
@@ -44,7 +44,7 @@ func TestGorillaMuxRouter(t *testing.T) {
 		})
 	})
 
-	t.Run("create swagger handler", func(t *testing.T) {
+	t.Run("create openapi handler", func(t *testing.T) {
 		handlerFunc := ar.SwaggerHandler("text/html", []byte("some data"))
 		echoRouter.GET("/oas", handlerFunc)